+++ Shared/ViewModels/SettingsVM.cs
@@ -2,8 +2,8 @@
 {
     public class SettingsVM
     {
-        public float HustleBenchmarkSeconds { get; set; }
-        public float OutlierSeconds { get; set; }
+        public int HustleBenchmarkSeconds { get; set; }
+        public int OutlierSeconds { get; set; }
         public int MinDels { get; set; }
         public int MinTrackedPercentage { get; set; }
         public int MaxOverspeeds { get; set; }
